+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/_client_factory.py
@@ -60,6 +60,10 @@ def cf_packet_capture(_):
     return _network_client_factory().packet_captures
 
 
+def cf_private_access(_):
+    return _network_client_factory().available_private_access_services
+
+
 def cf_public_ip_addresses(_):
     return _network_client_factory().public_ip_addresses