+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_client_factory.py
@@ -21,7 +21,7 @@ def cf_ni(_):
     return ni
 
 
-def cf_public_ip_addresses(_):
+def cf_public_ip_addresses():
     from azure.cli.core.profiles import ResourceType
     from azure.cli.core.commands.client_factory import get_mgmt_service_client
     public_ip_ops = get_mgmt_service_client(ResourceType.MGMT_NETWORK).public_ip_addresses