+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -4,6 +4,7 @@ Release History
 ===============
 (unreleased)
 +++++++++++++++++++
+* vm/vmss: support to assign identity w/o creating role assignments
 * vm: apply storage sku on attaching data disks
 
 2.0.11 (2017-07-27)