+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -2,8 +2,9 @@
 
 Release History
 ===============
-(unreleased)
+2.0.11 (2017-07-27)
 +++++++++++++++++++
+* vmss: support configuring nsg
 * vmss: fix a bug that dns server is not configured right.
 * vm/vmss: support managed service identity
 * `vmss create`: Fix issue where creating with existing load balancer required `--backend-pool-name`.